:recycle: refactor(server): Refactor code
