+++ apps/server/src/datasources/controllers/youtube/query.ts
@@ -1,11 +1,11 @@
 import { Request, Response } from 'express';
 
-import { searchVideo, getAudio, getVideoTrending } from '@utils/controllers';
+import { ytb } from '@utils/controllers';
 import { useHttpHandler } from '@utils/useHttpHandler';
 
 export const listSong = useHttpHandler(async (req: Request, res: Response): Promise<Response> => {
   const { search, token, apiKey } = req.body;
-  const results = await searchVideo(search, token, apiKey);
+  const results = await ytb.searchVideo(search, token, apiKey);
 
   return res.status(200).json({
     data: results,
@@ -14,7 +14,7 @@ export const listSong = useHttpHandler(async (req: Request, res: Response): Prom
 
 export const song = useHttpHandler(async (req: Request, res: Response): Promise<Response> => {
   const { yId } = req.params;
-  const result = await getAudio(yId);
+  const result = await ytb.getAudio(yId);
 
   return res.status(200).json({
     data: result,
@@ -22,7 +22,7 @@ export const song = useHttpHandler(async (req: Request, res: Response): Promise<
 });
 
 export const listSongTrending = useHttpHandler(async (req: Request, res: Response): Promise<Response> => {
-  const songs = await getVideoTrending();
+  const songs = await ytb.getVideoTrending();
 
   return res.status(200).json({
     data: { songs },
