+++ apps/server/src/datasources/models/index.ts
@@ -1,1 +1,1 @@
-export * from './user';
+export * from './User';
