+++ apps/server/src/datasources/utils/controllers/index.ts
@@ -1,22 +1,3 @@
-import { compareHash, generateHash } from './auth';
-import { createToken, verifyToken } from './token';
-import getAudio from './youtube/getAudio';
-import getVideoTrending from './youtube/getVideoTrending';
-import searchVideo from './youtube/searchVideo';
-
-export { searchVideo, getAudio, getVideoTrending };
-export const ytb = {
-  searchVideo,
-  getAudio,
-  getVideoTrending,
-};
-
-export const yeToken = {
-  createToken,
-  verifyToken,
-};
-
-export const authUtils = {
-  generateHash,
-  compareHash,
-};
+export * from './youtube';
+export * from './user';
+export * from './nodemailer';
