+++ apps/server/src/datasources/utils/controllers/user/token.ts
@@ -1,16 +1,6 @@
+import { TParamsCreateToken, TParamsVerifyToken } from '@utils/interface';
 import * as jwt from 'jsonwebtoken';
 import { jwtConfig } from 'src/config';
-
-type TParamsCreateToken = {
-  payload: string | object | Buffer;
-  expiration?: number;
-};
-
-type TParamsVerifyToken = {
-  token: string;
-  options?: jwt.VerifyOptions;
-};
-
 /**
  * It creates a token using the payload and expiration passed in as parameters
  * @param {TParamsCreateToken}  - TParamsCreateToken
