+++ apps/server/src/datasources/utils/interface/nodemailer.ts
@@ -1,0 +1,8 @@
+export type TParamsSendMail = {
+  from?: string;
+  to: string | string[];
+  subject: string;
+  text?: string;
+  html: string;
+  isUseTestAccount?: boolean;
+};
