+++ apps/server/src/datasources/utils/interface/user.ts
@@ -1,0 +1,11 @@
+import * as jwt from 'jsonwebtoken';
+
+export type TParamsCreateToken = {
+  payload: string | object | Buffer;
+  expiration?: number;
+};
+
+export type TParamsVerifyToken = {
+  token: string;
+  options?: jwt.VerifyOptions;
+};
