--- apps/server/src/datasources/utils/interface/youtube/index.ts
@@ -1,2 +1,0 @@
-export * from './youtube';
-export * from './formatVideo';
